Renamed gui_components to controllers
